+++ chapter-1/hours-within-year.rb
@@ -1,0 +1,16 @@
+puts 
+puts "# = = = = = = = = = = "
+puts "# Chapter 1"
+puts "# File: hours-within-year.rb"
+puts "# = = = = = = = = = = "
+puts
+
+puts "Constants:"
+puts "Year: 1"
+puts "Days: 365"
+puts "Months: 12"
+puts 
+
+puts "Hours in the year:"
+puts 365*24
+puts