+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - README.md
 - big-ruby-program.rb
 - chapter-1.rb [examples from chapter-1]
+- hours-within-year [todo from chapter-1]